pypi/pypi-aplikacja/app/requirements.txt
my_api_library @ https://__token__:123-@git.e-science.pl/api/v4/projects/1929/packages/pypi